+++ backend/backend_dragonbee/src/main/java/com/example/shopdragonbee/repository/HoaDonChiTietRepository.java
@@ -16,7 +16,7 @@ import java.util.List;
 @Repository
 public interface HoaDonChiTietRepository extends JpaRepository<HoaDonChiTiet, Integer> {
 
-    HoaDonChiTiet getHoaDonChiTietByHoaDonAndSanPhamChiTietAndDonGia(HoaDon hoaDon, SanPhamChiTiet sanPhamChiTiet, Float donGia);
+    HoaDonChiTiet getHoaDonChiTietByHoaDonAndSanPhamChiTietAndDonGia(HoaDon hoaDon, SanPhamChiTiet sanPhamChiTiet, Double donGia);
 
 //    @Query("SELECT hdct.sanPhamChiTiet.sanPham FROM HoaDonChiTiet hdct " +
 //            "WHERE hdct.hoaDon.trangThai = 'Hoàn thành' " +
